+++ xmake.lua
@@ -991,6 +991,7 @@ target("test", function()
             "test-bitvec-signal",
             "test-no-internal-clock",
             "test-handles",
+            "test-force-release-coalesce",
             "test-native-clock",
             "test-queue-waitable",
             "test-dpic",